Use IfNotPresent image pull policy in router tests
CI uses the current branch's tests but was pulling newer images.  In the
case of router-http-echo-server.yaml, the test data specified the
openshift/origin-base image with a command that required Perl, which is
missing from newer versions of that image.

This commit changes router test data to specify the IfNotPresent image pull
policy so that tests will use locally built images.

This commit fixes openshift/origin#19887.

https://github.com/openshift/origin/issues/19887