Prefer to delete doubled-up pods of a ReplicaSet
When scaling down a ReplicaSet, if a replica is doubled up on the same node
as a replica belonging to a related ReplicaSet, delete that pod first.
ReplicaSets are considered related if they have the same owner.

The intention of this change is a rolling update of a deployment should
prefer to scale down the old replicaset by deleting pods that are colocated
with pods of the new replicaset.

* vendor/k8s.io/kubernetes/pkg/controller/controller_utils.go:
(ActivePodsWithRanks): New type to sort pods using a given ranking.
* vendor/k8s.io/kubernetes/pkg/controller/replicaset/replica_set.go
(getReplicaSetsWithSameController): New method.  Given a ReplicaSet, return
all ReplicaSets that have the same owner.
(manageReplicas): Call getIndirectlyRelatedPods, and pass its result to
getPodsToDelete.
(getIndirectlyRelatedPods): New method.  Given a ReplicaSet, return all
pods that are owned by any ReplicaSet with the same owner.
(getPodsToDelete): Add an argument for related pods.  Use related pods and
activePodsWithLocalCounts to into account whether the pod is doubled up.
* vendor/k8s.io/kubernetes/pkg/controller/replicaset/replica_set_test.go
(TestGetPodsToDelete): Fix call to getPodsToDelete.